Update Block Breaker.rkt
I created the function "MAKE-STRUCTURE" which takes as argument a list "start" of exact 2 elements first being a number representing the starting point on x and second another number being the starting point on y. The arguments lines and columns must be non-negative integers, and greater than 0 to tell the function how many rows and columns we ask for. The function adds the blocks from left to right,  top to bottom. Best practice is to give a start point in top left corner, not too close to the edges. E.g: (MAKE-STRUCTURE `(75 125) 3 10) returns:
'(((75 125)(110 125)(145 125)(180 125)(215 125)(250 125)(285 125)(320 125)(355 125)(390 125))
  ((75 138)(110 138)(145 138)(180 138)(215 138)(250 138)(285 138)(320 138)(355 138)(390 138))
  ((75 151)(110 151)(145 151)(180 151)(215 151)(250 151)(285 151)(320 151)(355 151)(390 151)))

Where a nested list of 2 elements represents the position of a block.

Next step is to add the function that places the blocks into the main scene using this structure.

I have removed the manual defined blocks placement on the scene and their definition and I have defined "BLOCK" being the rectangle we are going to use.